% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfactor.R
\name{as.cfactor}
\alias{as.cfactor}
\alias{as.cfactor.default}
\alias{as.cfactor.factor}
\title{\code{cfactor} coercion}
\usage{
as.cfactor(x, codes = NULL, ordered = is.ordered(x))

\method{as.cfactor}{default}(x, codes = NULL, ordered = is.ordered(x))

\method{as.cfactor}{factor}(x, codes = NULL, ordered = is.ordered(x))
}
\arguments{
\item{x}{a vector of values representing labels for factor levels}

\item{codes}{named vector of unique codes that declares the mapping of labels
to codes}

\item{ordered}{logical flag to determine if the codes should be regarded as
ordered (in the order given).}
}
\value{
a new \code{cfactor}
}
\description{
Add codes to a vector of labels
}
\seealso{
Other cfactor utility functions: 
\code{\link{as.codes}()},
\code{\link{codes}()}
}
\concept{cfactor utility functions}
