% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-na.R
\name{as.na_collector}
\alias{as.na_collector}
\alias{as.na_collector.default}
\alias{as.na_collector.interlacer_na_collector}
\alias{as.na_collector.character}
\alias{as.na_collector.integer}
\alias{as.na_collector.double}
\title{Missing reason collector shortcuts}
\usage{
as.na_collector(x)

\method{as.na_collector}{default}(x)

\method{as.na_collector}{interlacer_na_collector}(x)

\method{as.na_collector}{character}(x)

\method{as.na_collector}{integer}(x)

\method{as.na_collector}{double}(x)
}
\arguments{
\item{x}{a value to convert into an \verb{na_col_*()} object}
}
\value{
a new missing reason collector object
}
\description{
TODO: Write me
}
\concept{missing reason collector shortcuts}
