% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-value.R
\name{as.value_collector}
\alias{as.value_collector}
\alias{as.value_collector.collector}
\alias{as.value_collector.collector_factor}
\alias{as.value_collector.interlacer_value_collector}
\alias{as.value_collector.character}
\title{Value collector shortcuts}
\usage{
as.value_collector(x)

\method{as.value_collector}{collector}(x)

\method{as.value_collector}{collector_factor}(x)

\method{as.value_collector}{interlacer_value_collector}(x)

\method{as.value_collector}{character}(x)
}
\arguments{
\item{x}{a value to convert into an \verb{v_col_*()} object}
}
\value{
a new value collector object
}
\description{
TODO: Write me
}
\concept{value collector shortcuts}
