% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-na.R, R/collectors-value.R,
%   R/collectors-x.R
\name{as.na_collector}
\alias{as.na_collector}
\alias{as.value_collector}
\alias{as.x_collector}
\title{Collector shortcuts}
\usage{
as.na_collector(x)

as.value_collector(x)

as.x_collector(x)
}
\arguments{
\item{x}{a value to convert into an extended collector, value collector,
or missing reason collector.}
}
\description{
TODO: Write me
}
\keyword{internal}
