% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_channels.R
\name{coalesce_channels}
\alias{coalesce_channels}
\title{Coalesce missing reasons in a data frame}
\usage{
coalesce_channels(
  x,
  default_reason = getOption("default_missing_reason"),
  keep = c("values", "missing")
)
}
\arguments{
\item{x}{A data frame}

\item{default_reason}{When a variable is missing a value and a missing
reason, the default missing reason to fill in.}

\item{keep}{When a variable has both a value and missing reason, choose which
to keep. (A properly formed deinterlaced data frame has values OR missing
reasons)}
}
\value{
A deinterlaced tibble.
}
\description{
Mutations of deinterlaced data frames can result in variables that either
have both values and missing reasons, or no values and no missing reasons.
`coalesce_channels()` takes care of both situations. In the case where
there is both a value and missing reason, it will choose which to keep based
on the `keep` paramter. In case where no value or missing reason exists, it
will fill the missing reason with the `default_reason` parameter.

Mutations can also create new value columns without companion missing reason
columns. In that case, a new missing reason will be created and filled with
`default_reason` wherever there are missing values in the value column. (
This behavior can also be used to stub missing reason columns for value-only
data frames)
}
