% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_missing_reasons.R
\name{coalesce_missing_reasons}
\alias{coalesce_missing_reasons}
\title{Coalesce missing reasons in a dataframe}
\usage{
coalesce_missing_reasons(
  x,
  keep = c("values", "missing"),
  default_reason = getOption("default_missing_reason")
)
}
\arguments{
\item{x}{A dataframe}

\item{keep}{When a variable has both a value and missing reason, choose which
to keep. (A properly formed deinterlaced dataframe has values OR missing
reasons)}

\item{default_reason}{When a variable is missing a value and a missing
reason, the default missing reason to fill in.}
}
\value{
A deinterlaced tibble.
}
\description{
Mutations of deinterlaced data frames can result in variables that either
have both values and missing reasons, or no values and no missing reasons.
`coalesce_missing_reasons()` takes care of both situations. In the case where
there is both a value and missing reason, it will choose which to keep based
on the `keep` paramter. In case where no value or missing reason exists, it
will fill the missing reason with the `default_reason` parameter.

Mutations can also create new value columns without companion missing reason
columns. In that case, a new missing reason will be created and filled with
`default_reason` wherever there are missing values in the value column. (
This behavior can also be used to stub missing reason columns for value-only
dataframes)
}
