% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfactor.R
\name{codes<-}
\alias{codes<-}
\title{Set the codes for a `cfactor``}
\usage{
codes(x) <- value
}
\arguments{
\item{value}{a named vector of codes for the \code{cfactor}}
}
\description{
Set the codes for a \code{cfactor}, similar to \verb{levels<-()}
}
\keyword{internal}
