% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfactor.R
\name{codes}
\alias{codes}
\alias{codes<-}
\alias{levels.interlacer_cfactor}
\title{\code{cfactor} attributes}
\usage{
codes(x, ...)

codes(x) <- value

\method{levels}{interlacer_cfactor}(x)
}
\arguments{
\item{x}{a \code{cfactor}}

\item{...}{additional arguments (not used)}
}
\value{
\code{levels()} returns the levels of the \code{cfactor} (as a vector of
character labels); \code{codes()} returns a named vector representing the codes
for the \code{cfactor}
}
\description{
Return the levels or codes of a \code{cfactor}
}
\keyword{internal}
