% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deinterlace_type_convert.R
\name{deinterlace_type_convert}
\alias{deinterlace_type_convert}
\title{Convert character columns and deinterlace missing reasons in existing data
frame}
\usage{
deinterlace_type_convert(x, col_types = NULL, na = c("", "NA"), ...)
}
\arguments{
\item{x}{A data frame}

\item{col_types}{One of `NULL`, a [readr::cols()] specification, or a string.}

\item{na}{Character vector of strings to interpret as missing values.}

\item{...}{additional parameters to pass to `readr::type_convert()`}
}
\value{
A [tibble()].with separate columns for values and missing reasons
for each variable.
}
\description{
This is a simple wrapper for `readr::type_convert()` that deinterlaces
missing reasons in addition to parsing values.
}
