% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_missing_reasons}
\alias{drop_missing_reasons}
\title{Drop missing reasons from a deinterlaced dataframe}
\usage{
drop_missing_reasons(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A tibble without missing reason columns.
}
\description{
Drop the missing reason columns from a deinterlaced dataframe, turning it
into a regular dataframe with unlabelled `NA` values.
}
