% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{flatten_channels}
\alias{flatten_channels}
\title{Flatten a \code{interlaced} vector}
\usage{
flatten_channels(x, ...)
}
\arguments{
\item{x}{An \code{interlaced} vector}

\item{...}{Additional arguments, not used}
}
\value{
The vector, flattened
}
\description{
\code{flatten_channels()} flattens an \code{interlaced} vector into a single channel.
This is useful as a step right before writing an \code{interlaced} vector to a
file, for example.
}
\seealso{
Other interlaced type functions: 
\code{\link{is.empty}()},
\code{\link{value_channel}()}
}
\concept{interlaced type functions}
