% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{\%==\%}
\alias{\%==\%}
\alias{\%!=\%}
\title{Multichannel comparison}
\usage{
x \%==\% y

x \%!=\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector to compare}
}
\value{
A boolean vector indicating the result of the comparison
}
\description{
Regular comparison operators (i.e. \code{==} and \code{!=}) only apply to value
channel; when used with missing values they always result in \code{NA}.
By contrast,
the multichannel comparison operators \verb{\%==\%} and \verb{\%!=\%} compare BOTH the
value and missing reason channels of a variable. Multichannel equality
\verb{\%==\%} operates identically to regular \code{==} when both operands are values,
but will return \code{FALSE} instead of \code{NA} when one is missing. When both values
being compared with \verb{\%==\%} are missing values, their missing reasons are
compared. Multichannel comparison operators always return boolean values,
never \code{NA}
}
