% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{interlaced}
\alias{interlaced}
\alias{as.interlaced}
\alias{as.interlaced.default}
\alias{as.interlaced.interlacer_interlaced}
\alias{as.interlaced.data.frame}
\alias{is.interlaced}
\title{Construct an \code{interlaced} vector}
\usage{
interlaced(x, na = NULL)

as.interlaced(x, na = NULL, ...)

\method{as.interlaced}{default}(x, na = NULL, ...)

\method{as.interlaced}{interlacer_interlaced}(x, ...)

\method{as.interlaced}{data.frame}(x, ...)

is.interlaced(x)
}
\arguments{
\item{x}{A vector or list of values}

\item{na}{A vector of values to interpret as missing values}

\item{...}{Additional arguments, not used}
}
\value{
An \code{interlaced} vector
}
\description{
The \code{interlaced} type extends vectors by adding a "missing reason" channel
which can be used to distinguish different types of missingness. The
\code{interlaced()} function constructs a new \code{interlaced} vector from a vector
or list of values.
}
