% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced_type_convert.R
\name{interlaced_type_convert}
\alias{interlaced_type_convert}
\title{Convert character columns into interlaced values and missing reasons}
\usage{
interlaced_type_convert(
  df,
  col_types = NULL,
  na = c("", "NA"),
  trim_ws = TRUE,
  locale = default_locale()
)
}
\arguments{
\item{col_types}{One of `NULL`, a [readr::cols()] specification, or a string.}

\item{na}{Character vector of strings to interpret as missing values.}

\item{x}{A data frame}

\item{...}{additional parameters to pass to `readr::type_convert()`}
}
\value{
A [tibble()] with interlaced columns
}
\description{
This is a simple wrapper for `readr::type_convert()` that converts character
columns into columns with interlaced values and missing reasons
}
