% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interlacer_progress}
\alias{interlacer_progress}
\title{Determine whether progress bars should be shown}
\usage{
interlacer_progress()
}
\description{
By default, interlacer shows progress bars. However, progress reporting is
suppressed if any of the following conditions hold:
- The bar is explicitly disabled by setting
  `options(interlacer.show_progress = FALSE)`.
- The code is run in a non-interactive session, as determined by
  [rlang::is_interactive()].
- The code is run in an RStudio notebook chunk, as determined by
  `getOption("rstudio.notebook.executing")`.
}
