% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interlacer_show_col_types}
\alias{interlacer_show_col_types}
\title{Determine whether column types should be shown}
\usage{
interlacer_show_col_types()
}
\description{
Wrapper around `getOption("interlacer.show_col_types")` that implements
some fall back logic if the option is unset. This returns:
* `TRUE` if the option is set to `TRUE`
* `FALSE` if the option is set to `FALSE`
* `FALSE` if the option is unset and we appear to be running tests
* `NULL` otherwise, in which case the caller determines whether to show
  column types based on context, e.g. whether `interlacer_show_col_types` or
  actual `col_types` were explicitly specified
}
