% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interlacer_threads}
\alias{interlacer_threads}
\title{Determine how many threads interlacer should use when processing. If not
running on a Unix-based system (like Windows), return 1 because they do not
support [parallel::mclapply()], which is what is currently being used for
parallel processing.}
\usage{
interlacer_threads()
}
\description{
The number of threads returned can be set by
- If `.Platform$OS.type != "unix"`, return `1`
- The global option `interlacer.num_threads`
- The environment variable `VROOM_THREADS`
- The value of [parallel::detectCores()]
}
