% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{is.empty}
\alias{is.empty}
\title{Test if a value is missing and lacks a missing reason}
\usage{
is.empty(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical vector the same length as x, containing TRUE for all
empty elements, and FALSE otherwise.
}
\description{
When a value is missing both a value and a missing reason, it is considered
"empty". \code{is.empty()} checks for these type of values. Regular \code{NA} values
(with no missing reasons) are also considered "empty".
}
\seealso{
Other interlaced type functions: 
\code{\link{flatten_channels}()},
\code{\link{value_channel}()}
}
\concept{interlaced type functions}
