% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{levels<-.interlacer_interlaced}
\alias{levels<-.interlacer_interlaced}
\alias{na_levels<-}
\title{Set the factor level attributes of \code{interlaced} vectors}
\usage{
\method{levels}{interlacer_interlaced}(x) <- value

na_levels(x) <- value
}
\arguments{
\item{value}{A vector of new levels to set}
}
\description{
Set the factor level attributes of \code{interlaced} vectors
}
\keyword{internal}
