% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{levels.interlacer_interlaced}
\alias{levels.interlacer_interlaced}
\alias{na_levels}
\alias{na_levels<-}
\title{Factor level attributes of \code{interlaced} vectors}
\usage{
\method{levels}{interlacer_interlaced}(x)

na_levels(x)

na_levels(x) <- value
}
\arguments{
\item{x}{an \code{interlaced} vector}
}
\value{
The levels of the values or missing reason channel
}
\description{
The base S3 \code{levels()} function is overloaded for \code{interlaced} vectors, so
when the value channel is a factor type, \code{levels()} will return its levels.
Similarly \code{na_levels()} will return the levels for the missing reason
channel.
}
\keyword{internal}
