% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{map_value_channel}
\alias{map_value_channel}
\alias{map_na_channel}
\title{\code{interlaced} functional utilities}
\usage{
map_value_channel(x, fn)

map_na_channel(x, fn)
}
\arguments{
\item{x}{an interlaced vector}

\item{fn}{a function that maps values or missing reasons to new values}
}
\value{
a new interlaced vector, modified according to the supplied function
}
\description{
\code{map_value_channel()} modifies the values of an \code{interlaced}
vector. \code{map_na_channel()} modifies the missing reason channel of an
\code{interlaced} vector.
}
\seealso{
Other interlaced type functions: 
\code{\link{flatten_channels}()},
\code{\link{is.empty}()},
\code{\link{value_channel}()}
}
\concept{interlaced type functions}
