% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{na}
\alias{na}
\title{Interpret a value as a missing reason}
\usage{
na(x = unspecified())
}
\arguments{
\item{x}{A character or numeric value}
}
\value{
An \code{interlaced} value
}
\description{
\code{na()} lifts a value into an \code{interlaced} missing reason channel.
}
