% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-na.R
\name{na_col_default}
\alias{na_col_default}
\alias{na_col_none}
\alias{na_col_integer}
\alias{na_col_factor}
\alias{na_col_cfactor}
\title{Missing reason collectors}
\usage{
na_col_default()

na_col_none()

na_col_integer(...)

na_col_factor(...)

na_col_cfactor(...)
}
\arguments{
\item{...}{values to interpret as missing values. In the case of
\code{na_col_cfactor()}, arguments must be named.}
}
\value{
a new missing reason collector object
}
\description{
TODO: Write me
}
\seealso{
Other collectors: 
\code{\link{v_col_guess}()},
\code{\link{x_col}()}
}
\concept{collectors}
