% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{na_spec}
\alias{na_spec}
\title{Examine the NA spec of a data frame}
\usage{
na_spec(x)
}
\arguments{
\item{x}{The data frame object to extract from}
}
\value{
An \code{na_col_spec} object
}
\description{
Like \code{readr::spec()}, \code{na_spec()} extracts the NA column specification from
a tibble created by \verb{read_interlaced_*}
}
