% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-interlaced.R
\name{parse_interlaced}
\alias{parse_interlaced}
\title{Parse a \code{character} vector into an \code{interlaced} vector type}
\usage{
parse_interlaced(x, na, .default = v_col_guess())
}
\arguments{
\item{x}{A character vector}

\item{na}{A missing reason collector (e.g. \code{na_col_integer()}), or one of
its shortcuts (e.g. a list of missing values)}

\item{.default}{A value collector to parse the character values (e.g.
\code{v_col_double()}, \code{v_col_integer()}, etc.)}
}
\value{
An \code{interlaced} vector
}
\description{
\code{parse_interlaced} converts a character vector to an \code{interlaced} vector
by parsing it with a readr \code{collector} type.
}
