% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_interlaced_delim}
\alias{read_interlaced_delim}
\alias{read_interlaced_csv}
\alias{read_interlaced_csv2}
\alias{read_interlaced_tsv}
\title{Read an delimited file with interlaced missing reasons into a tibble}
\usage{
read_interlaced_delim(
  file,
  delim = NULL,
  col_types = NULL,
  col_select = NULL,
  na = c("", "NA"),
  ...
)

read_interlaced_csv(
  file,
  col_types = NULL,
  col_select = NULL,
  na = c("", "NA"),
  ...
)

read_interlaced_csv2(
  file,
  col_types = NULL,
  col_select = NULL,
  na = c("", "NA"),
  ...
)

read_interlaced_tsv(
  file,
  col_types = NULL,
  col_select = NULL,
  na = c("", "NA"),
  ...
)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either
a single string or a raw vector).}

\item{delim}{Single character used to separate fields within a record.}

\item{col_types}{One of `NULL`, a [readr::cols()] specification, or a string. In
addition to the `col_*` specifiers provided by `readr`, `icol_*()`
specifiers may be used. See `vignette("interlacer")` for more details.}

\item{na}{Character vector of strings to interpret as missing values. These
values will become the factor levels of the missing reason column.}

\item{...}{Additional parameters to pass to `read_delim`}
}
\value{
A [tibble()].with separate columns for values and missing reasons
for each variable.
}
\description{
The `read_interlaced_*()`, family of functions extend `readr`'s
`read_delim()`, `read_csv`, etc. functions for use on data sources where
values are interlaced with missing reasons. These functions return a tibble
with two columns for each interlaced source column: a column with
values, and a column with missing reasons. Missing reason columns are named
by taking the value column name and surrounding it by dots
(e.g. missing reasons for "col_name" are read into a column named
".col_name.")
}
\examples{
# Beep boop
}
