% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-value.R
\name{v_col_guess}
\alias{v_col_guess}
\alias{v_col_cfactor}
\alias{v_col_character}
\alias{v_col_date}
\alias{v_col_datetime}
\alias{v_col_double}
\alias{v_col_factor}
\alias{v_col_integer}
\alias{v_col_big_integer}
\alias{v_col_logical}
\alias{v_col_number}
\alias{v_col_skip}
\alias{v_col_time}
\title{Value collectors}
\usage{
v_col_guess()

v_col_cfactor(codes, ordered = FALSE)

v_col_character()

v_col_date(format = "")

v_col_datetime(format = "")

v_col_double()

v_col_factor(levels = NULL, ordered = FALSE)

v_col_integer()

v_col_big_integer()

v_col_logical()

v_col_number()

v_col_skip()

v_col_time(format = "")
}
\arguments{
\item{codes}{A named vector of unique codes that declares the mapping of labels
to codes.}

\item{ordered}{Is it an ordered factor?}

\item{format}{A format specification, as described in \code{readr::col_datetime()}}

\item{levels}{Character vector of the allowed levels. When levels = \code{NULL}
(the default), levels are discovered from the unique values of x, in the
order in which they appear in x.}
}
\value{
a new value collector object
}
\description{
TODO: Write me
}
\seealso{
Other collectors: 
\code{\link{na_col_default}()},
\code{\link{x_col}()}
}
\concept{collectors}
