% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{value_channel}
\alias{value_channel}
\alias{na_channel}
\title{Access the channels of an \code{interlaced} vector}
\usage{
value_channel(x, ...)

na_channel(x, ...)
}
\arguments{
\item{x}{An \code{interlaced} vector}

\item{...}{Additional arguments, not used}
}
\value{
The value or missing reasons channel
}
\description{
\itemize{
\item \code{value_channel()} returns the value channel of an \code{interlaced} vector
\item \code{na_channel()} returns the missing reason channel of an \code{interlaced} vector
}
}
\seealso{
Other interlaced type functions: 
\code{\link{flatten_channels}()},
\code{\link{is.empty}()}
}
\concept{interlaced type functions}
