% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced-tidy-helpers.R
\name{where_value_channel}
\alias{where_value_channel}
\alias{where_na_channel}
\title{Select variables with a function applied on value or missing reason channels}
\usage{
where_value_channel(fn)

where_na_channel(fn)
}
\arguments{
\item{fn}{A function that returns \code{TRUE} or \code{FALSE} (technically, a
\emph{predicate} function). Can also be a purrr-like formula.}
}
\description{
\code{where_value_channel()} and \code{where_na_channel()} are simple wrappers for
\code{tidyselect::where()} that apply the selection function to the value or
missing reason channel of columns, respectively.
}
\concept{interlaced tidy helpers}
