% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-x.R
\name{x_col}
\alias{x_col}
\alias{as.x_collector}
\alias{as.x_collector.character}
\alias{as.x_collector.interlacer_x_collector}
\alias{as.x_collector.collector}
\alias{as.x_collector.interlacer_value_collector}
\title{Extended collectors}
\usage{
x_col(value_collector, na_collector = na_col_default())

as.x_collector(x)

\method{as.x_collector}{character}(x)

\method{as.x_collector}{interlacer_x_collector}(x)

\method{as.x_collector}{collector}(x)

\method{as.x_collector}{interlacer_value_collector}(x)
}
\arguments{
\item{value_collector}{a value collector}

\item{na_collector}{a missing reason collector}

\item{x}{a value to convert into an \code{x_col()} object}
}
\value{
a new extended collector object
}
\description{
TODO: Write me
}
\concept{extended collector constructors}
