% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-x-spec.R
\name{x_spec}
\alias{x_spec}
\title{Examine the extended column specifications for a data frame}
\usage{
x_spec(x)
}
\arguments{
\item{x}{a data frame loaded by \verb{read_interlaced_*()}}
}
\value{
An extended column specification object
}
\description{
TODO: Write me
}
