% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfactor.R
\name{as.codes}
\alias{as.codes}
\alias{as.codes.interlacer_interlaced}
\alias{as.codes.interlacer_cfactor}
\title{Convert a \code{cfactor} vector into a vector of its codes}
\usage{
as.codes(x, ...)

\method{as.codes}{interlacer_interlaced}(x, ...)

\method{as.codes}{interlacer_cfactor}(x, ...)
}
\arguments{
\item{x}{a \code{cfactor()}}

\item{...}{additional arguments (not used)}
}
\value{
a vector of coded values
}
\description{
TODO: Write this
}
