% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-x-spec.R
\name{as.x_col_spec}
\alias{as.x_col_spec}
\title{Extended column specification coercions}
\usage{
as.x_col_spec(x)
}
\arguments{
\item{x}{a value to coerce into an extended column specification}
}
\value{
an extended column specification
}
\description{
Coerce an object into a column specification. This is used internally to
parse the \code{col_types} argument in the \verb{read_interlaced_*()} family of
functions, so that it can accept a \code{readr::cols()} specification or \code{list()}.
}
\details{
It is an S3 function so that other packages may use the \code{col_types} argument
with their own custom objects.
}
\keyword{internal}
