% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-x-spec.R
\name{as.x_col_spec}
\alias{as.x_col_spec}
\title{Extended column specification coercions}
\usage{
as.x_col_spec(x)
}
\arguments{
\item{x}{a value to coerce into an extended column specification}
}
\value{
an extended column specification
}
\description{
TODO: Write me
}
\keyword{internal}
