% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_missing_cols}
\alias{drop_missing_cols}
\alias{drop_value_cols}
\title{Drop missing reasons from a deinterlaced dataframe}
\usage{
drop_missing_cols(x)

drop_value_cols(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A tibble without missing reason columns.
}
\description{
Drop the missing reason or value columns from a deinterlaced dataframe,
turning it into a regular dataframe with unlabelled `NA` values.
}
