% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{icol_logical}
\alias{icol_logical}
\alias{icol_integer}
\alias{icol_double}
\alias{icol_character}
\alias{icol_factor}
\alias{icol_date}
\alias{icol_time}
\alias{icol_datetime}
\alias{icol_number}
\title{Interlaced collectors for read_interlaced_*}
\usage{
icol_logical(na)

icol_integer(na)

icol_double(na)

icol_character(na)

icol_factor(na, levels = NULL, ordered = FALSE)

icol_date(na, format = "")

icol_time(na, format = "")

icol_datetime(na, format = "")

icol_number(na)
}
\arguments{
\item{na}{Character vector of strings to interpret as column-level missing
values}
}
\description{
Interlaced collector extend `readr` collector types (e.g. `col_double()`) to
allow column-level missing value specifications.
}
