% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlace_channels.R
\name{interlace_channels}
\alias{interlace_channels}
\title{Re-interlacce a deinterlaced dataframe}
\usage{
interlace_channels(x)
}
\arguments{
\item{x}{A deinterlaced dataframe}
}
\value{
An interlaced dataframe, that is, a dataframe with character columns
that contain both values and missing reasons.
}
\description{
This function will take a deinterlaced dataframe and re-interlace it by
combining value and misisng reason column pairs into single character
columns.
}
