% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{interlacer_example}
\alias{interlacer_example}
\title{Get path to interlacer example}
\usage{
interlacer_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
interlacer comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
interlacer_example()
interlacer_example("colors.csv")
}
