% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_names}
\alias{missing_names}
\alias{value_names}
\title{The names of an deinterlaced dataframe}
\usage{
missing_names(x)

value_names(x)
}
\arguments{
\item{x}{A deinterlaced dataframe}
}
\value{
A vector of missing reason or value column names.
}
\description{
Functions to get the names of missing reason columns or value columns in
an deinterlaced dataframe
}
