% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-na.R
\name{na_col_default}
\alias{na_col_default}
\alias{na_col_none}
\alias{na_col_integer}
\alias{na_col_factor}
\alias{na_col_cfactor}
\title{Missing reason collectors}
\usage{
na_col_default()

na_col_none()

na_col_integer(...)

na_col_factor(...)

na_col_cfactor(...)
}
\arguments{
\item{...}{values to interpret as missing values. In the case of
\code{na_col_cfactor()}, arguments must be named.}
}
\value{
a new missing reason collector object
}
\description{
Missing reason collectors are used in extended column specifications to
specify the type of a column's missing reason channel.
}
\details{
\code{na_col_default()} is used to signal that the missing reason type should
inherit the specification provided in the \verb{na = } argument of the
calling \verb{read_interlaced_*()} function
}
\seealso{
Other collectors: 
\code{\link{v_col_guess}()},
\code{\link{x_col}()}
}
\concept{collectors}
