% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{na_levels}
\alias{na_levels}
\alias{na_levels<-}
\alias{levels.interlacer_interlaced}
\alias{levels<-.interlacer_interlaced}
\title{Get the factor levels of the value or missing reason channel}
\usage{
na_levels(x)

na_levels(x) <- value

\method{levels}{interlacer_interlaced}(x)

\method{levels}{interlacer_interlaced}(x) <- value
}
\arguments{
\item{x}{an \code{interlaced} vector}

\item{value}{new levels to set}
}
\value{
The levels of the values or missing reason channel
}
\description{
The base S3 \code{levels()} function is overloaded for \code{interlaced} vectors, so
when the value channel is a factor type, \code{levels()} will return its levels.
Similarly \code{na_levels()} will return the levels for the missing reason
channel, when it is a \code{factor} type.
}
