% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaced.R
\name{parse_interlaced}
\alias{parse_interlaced}
\title{Parse a \code{character} vector into an \code{interlaced} vector type}
\usage{
parse_interlaced(x, na, .value_col = readr::col_guess())
}
\arguments{
\item{x}{A character vector}

\item{na}{A vector of values to interpret as missing values}

\item{.value_col}{A collector to parse the character values (e.g.
\code{readr::col_double()}, \code{readr::col_integer()}, etc.)}
}
\value{
An \code{interlaced} vector
}
\description{
\code{parse_interlaced} converts a character vector to an \code{interlaced} vector
by parsing it with a readr \code{collector} type.
}
