% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_interlaced_delim}
\alias{read_interlaced_delim}
\alias{read_interlaced_csv}
\alias{read_interlaced_csv2}
\alias{read_interlaced_tsv}
\alias{interlaced_vroom}
\title{Read an delimited file with interlaced missing reasons into a tibble}
\usage{
read_interlaced_delim(
  file,
  delim = NULL,
  quote = "\\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = readr::default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  comment = "",
  trim_ws = FALSE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  name_repair = "unique",
  num_threads = interlacer_threads(),
  progress = interlacer_progress(),
  show_col_types = interlacer_show_col_types(),
  skip_empty_rows = TRUE
)

read_interlaced_csv(
  file,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = readr::default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  name_repair = "unique",
  num_threads = interlacer_threads(),
  progress = interlacer_progress(),
  show_col_types = interlacer_show_col_types(),
  skip_empty_rows = TRUE
)

read_interlaced_csv2(
  file,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = readr::default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  name_repair = "unique",
  num_threads = interlacer_threads(),
  progress = interlacer_progress(),
  show_col_types = interlacer_show_col_types(),
  skip_empty_rows = TRUE
)

read_interlaced_tsv(
  file,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = readr::default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  name_repair = "unique",
  num_threads = interlacer_threads(),
  progress = interlacer_progress(),
  show_col_types = interlacer_show_col_types(),
  skip_empty_rows = TRUE
)

interlaced_vroom(
  file,
  delim = NULL,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  skip = 0,
  n_max = Inf,
  na = c("", "NA"),
  quote = "\\"",
  comment = "",
  skip_empty_rows = TRUE,
  trim_ws = TRUE,
  escape_double = TRUE,
  escape_backslash = FALSE,
  locale = vroom::default_locale(),
  guess_max = 100,
  num_threads = interlacer_threads(),
  progress = interlacer_progress(),
  show_col_types = interlacer_show_col_types(),
  .name_repair = "unique"
)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either
a single string or a raw vector).}

\item{delim}{Single character used to separate fields within a record.}

\item{col_types}{One of `NULL`, a [readr::cols()] specification, or a string. In
addition to the `col_*` specifiers provided by `readr`, `icol_*()`
specifiers may be used. See `vignette("interlacer")` for more details.}

\item{col_select}{Columns to include in the results. As with
[reader::read_delim], you can use the same mini-language as
[dplyr::select()] to refer to the columns by name.}

\item{na}{Character vector of strings to interpret as missing values. These
values will become the factor levels of the missing reason column.}

\item{...}{Additional parameters to pass to `read_delim`}
}
\value{
A deinterlaced [tibble()], that is, a tibble with separate columns
for values and missing reasonskfor each variable.
}
\description{
The `read_interlaced_*()`, family of functions extend `readr`'s
`read_delim()`, `read_csv`, etc. functions for use on data sources where
values are interlaced with missing reasons. These functions return a tibble
with two columns for each interlaced source column: a column with
values, and a column with missing reasons. Missing reason columns are named
by taking the value column name and surrounding it by dots
(e.g. missing reasons for "col_name" are read into a column named
".col_name.")
}
\examples{
# Beep boop
}
