% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_interlaced_delim}
\alias{write_interlaced_delim}
\alias{write_interlaced_csv}
\alias{write_interlaced_csv2}
\alias{write_interlaced_excel_csv}
\alias{write_interlaced_excel_csv2}
\alias{write_interlaced_tsv}
\title{Interlace a deinterlaced data frame and write it to a file}
\usage{
write_interlaced_delim(x, file, delim = " ", ...)

write_interlaced_csv(x, file, ...)

write_interlaced_csv2(x, file, ...)

write_interlaced_excel_csv(x, file, ...)

write_interlaced_excel_csv2(x, file, ...)

write_interlaced_tsv(x, file, ...)
}
\arguments{
\item{x}{A data frame or tibble to write to disk}

\item{file}{File or connection to write to}

\item{delim}{Delimiter used to separate values. Defaults to " " for
`write_interlaced_delim()`, "," for `write_interlaced_excel_csv()` and ";"
for `write_excel_csv2()`. Must be a single character.}

\item{...}{Additional parameters to pass to [readr]}
}
\value{
`write_interlaced_*` returns the input x invisibly
}
\description{
The `write_interlaced_*()` family of functions will take a deinterlaced
data frame, re-interlace it, and write it to a flie. The behavior of these
functions match their similarly named counterparts in [readr].
}
