% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-x.R
\name{x_col}
\alias{x_col}
\title{Construct an extended collector from a value and missing reason collector}
\usage{
x_col(value_collector, na_collector = na_col_default())
}
\arguments{
\item{value_collector}{a value collector}

\item{na_collector}{a missing reason collector}

\item{x}{a value to convert into an \code{x_col()} object}
}
\value{
a new extended collector object
}
\description{
TODO: Write me
}
\seealso{
Other collectors: 
\code{\link{na_col_default}()},
\code{\link{v_col_guess}()}
}
\concept{collectors}
