% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-x.R
\name{x_col}
\alias{x_col}
\title{Construct an extended collector for an extended column specification}
\usage{
x_col(value_collector, na_collector = na_col_default())
}
\arguments{
\item{value_collector}{a value collector}

\item{na_collector}{a missing reason collector}
}
\value{
a new extended collector object
}
\description{
Extended collectors are used in \code{x_cols()} column specifications to indicate
which value and missing reason channel types should be used when loading
data with \verb{read_interlaced_*()}.
}
\seealso{
Other collectors: 
\code{\link{na_collectors}},
\code{\link{value_collectors}}
}
\concept{collectors}
