% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors-x-spec.R
\name{x_cols}
\alias{x_cols}
\alias{x_cols_only}
\title{Extended column specifications}
\usage{
x_cols(..., .default = v_col_guess())

x_cols_only(...)
}
\arguments{
\item{...}{a named argument list of extended collector objects}

\item{.default}{a default value collector}
}
\value{
a new extended column specification
}
\description{
TODO: Write me
}
